move typechecker
