+++ src/eisen/adapters/new_vec.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from alpaca.concepts import Type
 from eisen.adapters.nodeinterface import AbstractNodeInterface
 from eisen.common.restriction import NewLetRestriction
-from eisen.validation.typeparser import TypeParser, TypeFactory
+from eisen.typecheck.typeparser import TypeParser, TypeFactory
 
 class NewVec(AbstractNodeInterface):
     ast_type = "new_vec"
